+++ src/crypto/CryptoNight_constants.h
@@ -57,19 +57,16 @@ constexpr const size_t   CRYPTONIGHT_HEAVY_MEMORY = 4 * 1024 * 1024;
 constexpr const uint32_t CRYPTONIGHT_HEAVY_MASK   = 0x3FFFF0;
 constexpr const uint32_t CRYPTONIGHT_HEAVY_ITER   = 0x40000;
 
-constexpr const size_t   CRYPTONIGHT_PICO_MEMORY = 256 * 1024;
-constexpr const uint32_t CRYPTONIGHT_PICO_MASK   = 0x1FFF0;
-constexpr const uint32_t CRYPTONIGHT_PICO_ITER   = 0x10000;
-
 constexpr const size_t   CRYPTONIGHT_FEMTO_MEMORY = 96 * 1024;
 constexpr const uint32_t CRYPTONIGHT_FEMTO_MASK   = 0xBFF0;
 constexpr const uint32_t CRYPTONIGHT_FEMTO_ITER   = 0x6000;
+
 
 template<Algo ALGO> inline constexpr size_t cn_select_memory()           { return 0; }
 template<> inline constexpr size_t cn_select_memory<CRYPTONIGHT>()       { return CRYPTONIGHT_MEMORY; }
 template<> inline constexpr size_t cn_select_memory<CRYPTONIGHT_LITE>()  { return CRYPTONIGHT_LITE_MEMORY; }
 template<> inline constexpr size_t cn_select_memory<CRYPTONIGHT_HEAVY>() { return CRYPTONIGHT_HEAVY_MEMORY; }
-template<> inline constexpr size_t cn_select_memory<CRYPTONIGHT_PICO>()  { return CRYPTONIGHT_PICO_MEMORY; }
+template<> inline constexpr size_t cn_select_memory<CRYPTONIGHT_FEMTO>()  { return CRYPTONIGHT_FEMTO_MEMORY; }
 
 
 inline size_t cn_select_memory(Algo algorithm)
@@ -85,8 +82,8 @@ inline size_t cn_select_memory(Algo algorithm)
     case CRYPTONIGHT_HEAVY:
         return CRYPTONIGHT_HEAVY_MEMORY;
 
-    case CRYPTONIGHT_PICO:
-        return CRYPTONIGHT_PICO_MEMORY;
+    case CRYPTONIGHT_FEMTO:
+        return CRYPTONIGHT_FEMTO_MEMORY;
 
     default:
         break;
@@ -100,7 +97,7 @@ template<Algo ALGO> inline constexpr uint32_t cn_select_mask()           { retur
 template<> inline constexpr uint32_t cn_select_mask<CRYPTONIGHT>()       { return CRYPTONIGHT_MASK; }
 template<> inline constexpr uint32_t cn_select_mask<CRYPTONIGHT_LITE>()  { return CRYPTONIGHT_LITE_MASK; }
 template<> inline constexpr uint32_t cn_select_mask<CRYPTONIGHT_HEAVY>() { return CRYPTONIGHT_HEAVY_MASK; }
-template<> inline constexpr uint32_t cn_select_mask<CRYPTONIGHT_PICO>()  { return CRYPTONIGHT_PICO_MASK; }
+template<> inline constexpr uint32_t cn_select_mask<CRYPTONIGHT_FEMTO>()  { return CRYPTONIGHT_FEMTO_MASK; }
 
 
 inline uint32_t cn_select_mask(Algo algorithm)
@@ -116,8 +113,8 @@ inline uint32_t cn_select_mask(Algo algorithm)
     case CRYPTONIGHT_HEAVY:
         return CRYPTONIGHT_HEAVY_MASK;
 
-    case CRYPTONIGHT_PICO:
-        return CRYPTONIGHT_PICO_MASK;
+    case CRYPTONIGHT_FEMTO:
+        return CRYPTONIGHT_FEMTO_MASK;
 
     default:
         break;
@@ -147,7 +144,7 @@ template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_LITE, VARIANT_1>
 template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_HEAVY, VARIANT_0>()    { return CRYPTONIGHT_HEAVY_ITER; }
 template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_HEAVY, VARIANT_XHV>()  { return CRYPTONIGHT_HEAVY_ITER; }
 template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_HEAVY, VARIANT_TUBE>() { return CRYPTONIGHT_HEAVY_ITER; }
-template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_PICO, VARIANT_TRTL>()  { return CRYPTONIGHT_PICO_ITER; }
+template<> inline constexpr uint32_t cn_select_iter<CRYPTONIGHT_FEMTO, VARIANT_CAT>()  { return CRYPTONIGHT_FEMTO_ITER; }
 
 
 inline uint32_t cn_select_iter(Algo algorithm, Variant variant)
@@ -164,8 +161,8 @@ inline uint32_t cn_select_iter(Algo algorithm, Variant variant)
     case VARIANT_DOUBLE:
         return CRYPTONIGHT_XAO_ITER;
 
-    case VARIANT_TRTL:
-        return CRYPTONIGHT_PICO_ITER;
+    case VARIANT_CAT:
+        return CRYPTONIGHT_FEMTO_ITER;
 
     case VARIANT_RWZ:
     case VARIANT_ZLS:
@@ -186,8 +183,8 @@ inline uint32_t cn_select_iter(Algo algorithm, Variant variant)
     case CRYPTONIGHT_HEAVY:
         return CRYPTONIGHT_HEAVY_ITER;
 
-    case CRYPTONIGHT_PICO:
-        return CRYPTONIGHT_PICO_ITER;
+    case CRYPTONIGHT_FEMTO:
+        return CRYPTONIGHT_FEMTO_ITER;
 
     default:
         break;
@@ -208,7 +205,7 @@ template<> inline constexpr Variant cn_base_variant<VARIANT_XAO>()    { return V
 template<> inline constexpr Variant cn_base_variant<VARIANT_RTO>()    { return VARIANT_1; }
 template<> inline constexpr Variant cn_base_variant<VARIANT_2>()      { return VARIANT_2; }
 template<> inline constexpr Variant cn_base_variant<VARIANT_HALF>()   { return VARIANT_2; }
-template<> inline constexpr Variant cn_base_variant<VARIANT_TRTL>()   { return VARIANT_2; }
+template<> inline constexpr Variant cn_base_variant<VARIANT_CAT>()   { return VARIANT_2; }
 template<> inline constexpr Variant cn_base_variant<VARIANT_GPU>()    { return VARIANT_GPU; }
 template<> inline constexpr Variant cn_base_variant<VARIANT_WOW>()    { return VARIANT_2; }
 template<> inline constexpr Variant cn_base_variant<VARIANT_4>()      { return VARIANT_2; }