+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 # v2.13.0
 - **[#938](https://github.com/xmrig/xmrig/issues/938) Added support for new algorithm `cryptonight/r`, short alias `cn/r` (also known as CryptoNightR or CryptoNight variant 4), for upcoming [Monero](https://www.getmonero.org/) fork on March 9, thanks [@SChernykh](https://github.com/SChernykh).**
 - [#939](https://github.com/xmrig/xmrig/issues/939) Added support for dynamic (runtime) pools reload.
-- [#932](https://github.com/xmrig/xmrig/issues/932) Fixed `cn-pico` hashrate drop, regression since v2.11.0.
+- [#932](https://github.com/xmrig/xmrig/issues/932) Fixed `cn-femto` hashrate drop, regression since v2.11.0.
 
 # v2.12.0
 - [#929](https://github.com/xmrig/xmrig/pull/929) Added support for new algorithm `cryptonight/wow`, short alias `cn/wow` (also known as CryptonightR), for upcoming [Wownero](http://wownero.org) fork on February 14.
@@ -28,7 +28,7 @@
 - [#292](https://github.com/xmrig/xmrig/issues/292) Fixed build on ARMv8 platforms if compiler not support hardware AES.
 
 # v2.10.0
-- [#904](https://github.com/xmrig/xmrig/issues/904) Added new algorithm `cn-pico/trtl` (aliases `cryptonight-turtle`, `cn-trtl`) for upcoming TurtleCoin (TRTL) fork.
+- [#904](https://github.com/xmrig/xmrig/issues/904) Added new algorithm `cn-femto/cat` (aliases `cryptonight-catalans`, `cn-cat`) for upcoming CatalansCoin (CAT) fork.
 - Default value for option `max-cpu-usage` changed to `100` also this option now deprecated.
 
 # v2.9.4